test(frontend/copilot): address coderabbit nitpicks on streaming tests
- Extract StreamHarness + click helpers into colocated copilotStream.harness.tsx
  (test file drops from 552 → 423 lines; harness is now reusable for future
  CoPilot SSE test files).
- Hoist userEvent to a static top-level import in the harness file (was
  dynamically imported per call from clickSend/clickStop).
- Drop vi.useRealTimers() from all afterEach blocks — no test in this file
  uses vi.useFakeTimers(), so restoring real timers was defensive noise and
  contradicted the project's documented "only use after fake timers" rule.
